refactor of auth routers callback functions exported to a controllers file + added validation of input data to endpoints register and login
