+++ src/api/RouteTable.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express'
-import AuthRouter from './routers/auth'
-import MiddlewareRouter from './routers/middleware'
+import AuthRouter from './routers/auth-router'
+import MiddlewareRouter from './routers/middleware-router'
 
 // The route table job comes once we are inside the application through the initial /api path.
 // Works as a gateway door to internal services, where each provided router holds a set of
